requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google-generativeai
openai